fix tests and bidding data, fix src refactor, update publish workflow
