+++ src/nemosis/defaults.py
@@ -56,9 +56,9 @@ table_types = {
     "SPDREGIONCONSTRAINT": "MMS",
     "SPDCONNECTIONPOINTCONSTRAINT": "MMS",
     "SPDINTERCONNECTORCONSTRAINT": "MMS",
-    "BIDPEROFFER_D": "BIDDING",
+    "BIDPEROFFER_D": "MMS",
     "DISPATCHINTERCONNECTORRES": "MMS",
-    "BIDDAYOFFER_D": "BIDDING",
+    "BIDDAYOFFER_D": "MMS",
     "DISPATCHREGIONSUM": "MMS",
     "FCAS_4_SECOND": "FCAS",
     "ELEMENTS_FCAS_4_SECOND": "STATIC",
